+++ packages/runtime/src/widgets/Form/FormItem.tsx
@@ -1,9 +1,9 @@
 import type { FormItemProps } from "antd";
 import { Form as AntForm } from "antd";
-import type { FormInputProps } from "./types";
 import { isString } from "lodash-es";
-import { EnsembleRuntime } from "../../runtime";
 import { unwrapWidget } from "@ensembleui/react-framework";
+import { EnsembleRuntime } from "../../runtime";
+import type { FormInputProps } from "./types";
 
 export type EnsembleFormItemProps<T> = FormItemProps & {
   values?: FormInputProps<T>;
@@ -17,6 +17,8 @@ export const EnsembleFormItem: React.FC<EnsembleFormItemProps<unknown>> = (
 ) => {
   const { values, ...rest } = props;
   const { backgroundColor: _, ...formItemStyles } = values?.styles ?? {};
+  const formInstance = AntForm.useFormInstance();
+
   return (
     <AntForm.Item
       className={values?.styles?.names}
@@ -36,7 +38,7 @@ export const EnsembleFormItem: React.FC<EnsembleFormItemProps<unknown>> = (
           </label>
         ) : null
       }
-      name={values?.id ?? values?.label}
+      name={formInstance ? values?.id ?? values?.label : undefined}
       rules={[{ required: Boolean(values?.required) }]}
       style={{
         margin: "0px",